feat: submit the pyment

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -69,9 +69,23 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
 
     this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
       this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const  navigationExtras: NavigationExtras = {state: order};
-      this.router.navigate(['checkout/success'], navigationExtras);
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if (result.paymentIntent) {
+          this.basketService.deleteLocalBasket(basket.id);
+          const  navigationExtras: NavigationExtras = {state: order};
+          this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error('Payment error');
+        }
+      });
     }, error => {
       this.toastr.error(error.message);
       console.log(error);
